refactor-#35: move activity entity to activity
