+++ src/catalog/activity/commands/handlers/create-activity.handler.ts
@@ -4,7 +4,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { RpcExceptionService } from '../../../../utils/exception-handling';
 import { ErrorValidationService } from '../../../../utils/error-validation';
 import { ActivityRepository } from '../../repositories';
-import { ActivityEntity } from '../../../../db/entities/activity.entity';
+import { ActivityEntity } from '../../entities';
 import { CreateActivityCommand } from '../impl';
 import { EsService } from '../../../../es/es.service';
 import { ISearchBody } from 'src/es/interfaces';
