+++ src/catalog/activity/entities/activity.entity.ts
