+++ src/catalog/activity/entities/index.ts
@@ -1,0 +1,1 @@
+export * from './activity.entity';
