+++ src/db/entities/index.ts
@@ -1,10 +1,8 @@
-import { ActivityEntity } from './activity.entity';
 import { CarsEntity } from './cars.entity';
 import { FlightEntity } from './flight.entity';
 import { HotelEntity } from './hotel.entity';
 
 export const CatalogEntities = [
-  ActivityEntity,
   CarsEntity,
   FlightEntity,
   HotelEntity,
